Move all fragment-related classes into dedicated file
